cache: simplify entry code
Consolidate the entry allocation paths using constants instead of
build tags.
