--- internal/cache/cgo_enabled.go
@@ -1,10 +1,0 @@
-// Copyright 2020 The LevelDB-Go and Pebble Authors. All rights reserved. Use
-// of this source code is governed by a BSD-style license that can be found in
-// the LICENSE file.
-
-//go:build cgo
-// +build cgo
-
-package cache
-
-const cgoEnabled = true
